requirements.txt
pytest
SQLAlchemy
psycopg2-binary
python-decouple
